backend/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
SQLAlchemy
psycopg2-binary
python-dotenv
